Light verb is now predicted (#33622)
Fix
